+++ bee-pow/src/score.rs
@@ -5,7 +5,9 @@
 
 use bee_ternary::{b1t6, Btrit, T1B1Buf, TritBuf, Trits, T1B1};
 
-use bee_crypto::ternary::sponge::{Sponge, UnrolledCurlP81};
+#[allow(deprecated)]
+use bee_crypto::ternary::sponge::{Sponge};
+use bee_crypto::ternary::sponge::{UnrolledCurlP81};
 use crypto::hashes::{blake2b::Blake2b256, ternary::HASH_LENGTH, Digest};
 
 /// Encapsulates the different steps that are used for scoring Proof of Work.
@@ -51,6 +53,7 @@ impl PoWScorer {
         self.pow_input.push(Btrit::Zero);
 
         // TODO: Consider using an output buffer here, for example by using the `Sponge` mechanism?
+        #[allow(deprecated)]
         self.curl.digest(self.pow_input.as_slice()).unwrap()
     }