+++ bee-crypto/CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Deprecated
 
 - Deprecated `CurlP` from `bee-crypto` (now uses `iota-crypto` instead)
+- Deprecated `Sponge` and `SpongeKind`
 
 ### Removed